Fix clustered Readme file (#1)
* Rename README.md to Redistricting.md

This file should not be the main Readme to this project. It offers information on Redistricting and will be named that

* Create README.md

Splitting of clustered readme: 
Created links for Vision and Stakeholder documents making them easier to find. 
Pointed to license, developmental conventions, and information on redistricting. 
Added a place for "Getting Started" documentation.

* Add files via upload

Better color scheme and resolution images for redistricting.md document

* Add files via upload

Removed spaces in file names

* Delete New figure A.png

* Delete New figure B.png

* Delete New figure C.png

* Delete New figure D.png

* Create NewRedistricting.md

Redistricting file updated to a more readable file format. Product description moved to the top, colors updated on images to make more accessible for users (followed colorblind guidelines). Overall flow easier to read and better to understand

* Rename Redistricting.md to OldRedistricting.md

* Update README.md

Pushed changes to redistricting document to readme

* Add files via upload

Getting Started File added to repository

* Rename Getting Started.md to GettingStarted.md

* Update GettingStarted.md